python_app/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
streamlit==1.26.0
prefect==2.13.0
pydantic<2.0.0,>=1.10.0
python-dotenv==1.0.0
aiohttp==3.8.5
redis==4.6.0
imap-tools==1.0.0
email-validator==2.0.0
jinja2==3.1.2
httpx==0.24.1
python-multipart==0.0.6
watchfiles==0.20.0
